Parse.com batch delete
This is very handy for batch deleting a number of objects. Using Regex,
you can type in a string of information your data may contain. The
Python will query the regex (ie fine all items that start with "x") and
then return json. The second "phase" of the file is to delete the object
if it appeared in the JSON results above. This is why the API has the be
called twice.

Use the "Results_Only.py" to simply see which objects are returned in
the first half of the code (if the json it out  puts). By doing this,
you can safely check which objects (rows) it will delete.

Don't forget to define your class, and of course fill our your API keys
from Parse.com.